Refactored demo.py to show how to use pyqtconfig from outside the package
